Use ios-deploy instead of libimobiledevice
Added an option to use ios-deploy instead of libimobiledevice.
It can be enabled by providing the path to the ios-deploy binary.

Also altered the display module to use provided streamWidth and streamHeight instead of pulling a screenshot.

Changed license

In so much as some of the content was created originally under an Apache License, that original content shall remain Apache License. In that the majority of the content has been altered and many new files have been created since that fork, all of that new content shall be under the new license from this point on.

As a result, am simply changing the license overall since the majority of the content has been changed since the fork.

Should this license change be found to be invalid, the licensing change shall still apply to all changes that are legally possible and all such content may not be used by the restricted parties as described in the new license regardless.